+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## 0.14.10
+### Updates
+- Undefined are note sent from the client.
+- Removed socket.io cookie by default.
 
 ## 0.14.9
 ### Features